push final

+++ ft_strlen.c
@@ -11,9 +11,11 @@
 /*                                                        /                   */
 /* ************************************************************************** */
 
-int	ft_strlen(char *str)
+#include "libft.h"
+
+size_t	ft_strlen(char *str)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
